+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added "Curse" effect; temporarily preventing the streamer from using any items.
 - Changed trap effects to retry if the player is already in the trap.
 - Changed project file layout.
+- Fixed crash during world save caused by incorrect cloning of ItemOwner Global Item.
 
 ## [1.0.9] - 2022-12-11