+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+# [1.1.0] - 2023-01-30
+
 - Added "Curse" effect; temporarily preventing the streamer from using any items.
 - Changed trap effects to retry if the player is already in the trap.
 - Changed project file layout.